requirements.txt
streamlit==1.30.0
beautifulsoup4==4.12.3
ollama
pyperclip==1.8.2
requests==2.31.0
tenacity==8.5.0
icalendar==5.0.12
openai==0.27.0

# For searching hotels and writing to Google Calendar:
google-api-python-client==2.92.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
